fix compile warning
git-svn-id: http://192.168.20.11/sequoiadb/trunk@24986 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/util/utilCache.cpp
@@ -298,7 +298,7 @@ namespace engine
       ossGetCurrentTime( t ) ;
       _lastTime = t.time * 1000 + t.microtm / 1000 ;
 
-      if ( ~0 == _length || offset < _start )
+      if ( (UINT32)~0 == _length || offset < _start )
       {
          _start = offset ;
       }
